+++ components/auth/legalModal.tsx
@@ -1,0 +1,30 @@
+'use client';
+
+import { Button } from '@nextui-org/button';
+import { Modal, ModalBody, ModalContent, ModalFooter, ModalHeader } from '@nextui-org/modal';
+
+interface LegalModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  title: string;
+  content: string;
+}
+
+export const LegalModal: React.FC<LegalModalProps> = ({ isOpen, onClose, title, content }) => {
+  return (
+    <Modal isOpen={isOpen} onClose={onClose} scrollBehavior='inside'>
+      <ModalContent>
+        <ModalHeader className="flex flex-col gap-1">{title}</ModalHeader>
+        <ModalBody className="text-sm">
+        {/* biome-ignore lint/security/noDangerouslySetInnerHtml: <explanation> */}
+        <div dangerouslySetInnerHTML={{ __html: content.replace(/\n/g, '<br>') }} />
+        </ModalBody>
+        <ModalFooter>
+          <Button color="danger" variant="light" onPress={onClose}>
+            Close
+          </Button>
+        </ModalFooter>
+      </ModalContent>
+    </Modal>
+  );
+};